Use setUpClass() and tearDownClass() in test_multiprocessing.
Each manager test class now uses a separate manager.  Also, process
pools are no longer created before starting any tests.

Note that warnings are written if the manager for a test case still
has live objects when it is shutdown.  This is true for a few test cases
which fail to wait for all child processes to end.